fixed booking

+++ db/migrate/20210220114207_add_total_cost_to_bookings.rb
@@ -1,5 +1,5 @@
 class AddTotalCostToBookings < ActiveRecord::Migration[6.0]
   def change
-    add_column :bookings, :total_cost, :decimal, precision: 2
+    add_column :bookings, :total_cost, :float 
   end
 end
